Move event-service to src
