+++ services/psibase_tests/CMakeLists.txt
@@ -13,7 +13,7 @@ endfunction()
 function(add suffix)
     service(TestKV "${suffix}" src/TestKV.cpp)
     service(TestTable "${suffix}" src/TestTable.cpp)
-    service(event-service "${suffix}" event-service.cpp)
+    service(EmitEvents "${suffix}" src/EmitEvents.cpp)
     service(memo-service "${suffix}" memo-service.cpp)
     service(clock-service "${suffix}" clock-service.cpp)
     service(subjective-service "${suffix}" subjective-service.cpp)
